+++ tests/integration-tests/tests/common/networking/security_groups.py
@@ -23,12 +23,12 @@ def delete_security_group(region: str, security_group_id: str):
 def describe_security_groups_for_network_interface(region: str, network_interface_id: str):
     logging.info(f"Describing Security Groups for Network Interface {network_interface_id}")
     try:
-        network_inyterface_description = _ec2(region).describe_network_interfaces(
+        network_interface_description = _ec2(region).describe_network_interfaces(
             NetworkInterfaceIds=[network_interface_id]
         )
         return [
             security_group["GroupId"]
-            for security_group in network_inyterface_description["NetworkInterfaces"][0]["Groups"]
+            for security_group in network_interface_description["NetworkInterfaces"][0]["Groups"]
         ]
     except Exception as e:
         logging.error(f"Cannot describe Security Groups for Network Interface {network_interface_id}: {e}")